+++ src/components/functional/toasts-list/components/custom-toast/components/custom-create-toast/custom-toast.tsx
@@ -32,9 +32,15 @@ export class CustomToast {
   }
 
   private initializeToast(container: HTMLElement) {
-    if (!container || container.hasChildNodes()) {
+    if (!container) {
       return;
     }
+
+    // Clear previous toast element
+    if (container.hasChildNodes()) {
+      container.innerHTML = '';
+    }
+
     const customElement = this.toast.instantiateToastElement();
     if (!customElement) {
       return;